version 1.25-1

+++ src/Swave.h
@@ -7,6 +7,7 @@
 #include <string.h>
 #include <stdlib.h>
 #include <math.h>
+#include "R.h"
 
 #ifndef Macintosh
 #include <sys/file.h>
@@ -49,8 +50,8 @@ typedef struct
   int resoln;
   int x; 
   int y; 
-  float W1f; 
-  float W2f; 
+  double W1f; 
+  double W2f; 
 } image_ext;
 */
 
@@ -60,7 +61,7 @@ extern int nint();
 extern double log2();
 extern double exp2();
 extern int iexp2();
-extern float fexp2();
+extern double fexp2();
 extern int find2power();
 
 /* uniform random number generator */
